+++ src/marker.cpp
@@ -10,7 +10,7 @@ visualization_msgs::Marker createSphereMarker(
 	double radius,               ///< The radius of the sphere
 	std::string ns,              ///< The namespace of the marker
 	ros::Duration lifetime,      ///< The lifetime of the marker
-	std::array<double, 4> color, ///< The color of the sphere in RGBA
+	std::array<float, 4> color,  ///< The color of the sphere in RGBA
 	int id                       ///< The id of this marker
 ) {
 	visualization_msgs::Marker marker;
@@ -43,7 +43,7 @@ visualization_msgs::Marker createCylinderMarker(
 	double height,               ///< The height of the cylinder
 	std::string ns,              ///< The namespace of the marker
 	ros::Duration lifetime,      ///< The lifetime of the marker
-	std::array<double, 4> color, ///< The color of the cylinder in RGBA
+	std::array<float, 4> color,  ///< The color of the cylinder in RGBA
 	int id                       ///< The id of this marker
 ) {
 	visualization_msgs::Marker marker;
@@ -72,7 +72,7 @@ visualization_msgs::Marker createBoxMarker(
 	Eigen::AlignedBox3d box,     ///< The position of the marker in frame_id
 	std::string ns,              ///< The namespace of the marker
 	ros::Duration lifetime,      ///< The lifetime of the marker
-	std::array<double, 4> color, ///< The color of the sphere in RGBA
+	std::array<float, 4> color,  ///< The color of the sphere in RGBA
 	int id                       ///< The id of this marker
 ) {
 	visualization_msgs::Marker marker;
